Add javadoc and change return type
Originally I had this just return void thinking that I would print out
what was going on.

searchForManaClumps will search for the biggest mana
clump and return it's size. I think I'm gonna refactor this.

analyzeDeck needs a score. I think I have figured out a good way to
determine if a deck is randomized enough. I will find the biggest
mana clump and compare it to the amount of mana in the deck returning
a double that will represent a percentage. This works well enough but
I would like to also figure out a way to repersent all the mana in the
deck not just the biggest mana clump.